[format] Fix formatting in cmake.
