+++ lib/Optimizer/CMakeLists.txt
@@ -1,9 +1,9 @@
 
 add_library(Optimizer
-            IROptimizer.cpp
-            GraphOptimizer.cpp
-            Lower.cpp
-            Quantization.cpp)
+              IROptimizer.cpp
+              GraphOptimizer.cpp
+              Lower.cpp
+              Quantization.cpp)
 
 target_link_libraries(Optimizer
                       PRIVATE
