+++ lib/Quantization/Base/Base.cpp
@@ -1,0 +1,34 @@
+/**
+ * Copyright (c) 2017-present, Facebook, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+#include "glow/Quantization/Base/Base.h"
+
+#include <cmath>
+
+namespace glow {
+namespace quantization {
+
+int8_t quantize(float input, const TensorQuantizationParams &TQP) {
+  float result = input / TQP.scale_ + TQP.offset_;
+  return quantization::clip<int32_t, int8_t>((int32_t)nearbyintf(result));
+}
+
+float dequantize(int8_t input, const TensorQuantizationParams &TQP) {
+  return TQP.scale_ * (input - TQP.offset_);
+}
+
+}
+}
