+++ lib/Quantization/Base/CMakeLists.txt
@@ -1,0 +1,6 @@
+add_library(QuantizationBase
+		QuantizationBase.cpp)
+
+target_link_libraries(QuantizationBase
+                      PUBLIC
+                        Support)
